combat "mouse enhancement"-induced jitteriness and unpredictability clamp linear acceleration relative to tank speed to 0.05 seconds from "full stop" to "full forward" on both the client and the server. server noisily drops packets where the acceleration is out of bounds client uses the same code as an inertia setting would to distribute the acceleration between multiple updates. needs testing and tuning.  for reference, the max I've been able to pull with fairly aggressive mouse settings is about 0.20 seconds from stop to _tankSpeed.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@11586 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/common/TextUtils.cxx
@@ -218,7 +218,7 @@ namespace TextUtils
       return false;
 
     int t = 0;
-    int len = strlen(duration);
+    int len = (int)strlen(duration);
     for (int i = 0; i < len; i++) {
       if (isdigit(duration[i])) {
 	t = t * 10 + (duration[i] - '0');
